# mobile/docs/apis/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Fridge2Fork API",
      "description": "냉장고 재료 기반 한식 레시피 추천 API",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/",
        "description": "Default server"
      },
      {
        "url": "/fridge2fork",
        "description": "Kubernetes ingress path"
      }
    ],
    "paths": {
      "/fridge2fork/v1/recipes/": {
        "get": {
          "tags": [
            "레시피"
          ],
          "summary": "Get Recipes",
          "description": "레시피 목록 조회 (scrape 마이그레이션 스키마 기반)",
          "operationId": "get_recipes_fridge2fork_v1_recipes__get",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": 1,
                "description": "페이지 번호",
                "default": 1,
                "title": "Page"
              },
              "description": "페이지 번호"
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "maximum": 100,
                "minimum": 1,
                "description": "페이지 크기",
                "default": 10,
                "title": "Size"
              },
              "description": "페이지 크기"
            },
            {
              "name": "search",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "검색어",
                "title": "Search"
              },
              "description": "검색어"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "title": "Response Get Recipes Fridge2Fork V1 Recipes  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/fridge2fork/v1/recipes/{rcp_sno}": {
        "get": {
          "tags": [
            "레시피"
          ],
          "summary": "Get Recipe",
          "description": "특정 레시피 상세 조회",
          "operationId": "get_recipe_fridge2fork_v1_recipes__rcp_sno__get",
          "parameters": [
            {
              "name": "rcp_sno",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Rcp Sno"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "title": "Response Get Recipe Fridge2Fork V1 Recipes  Rcp Sno  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      }
    }
  }